frontend: fix; don't display full data in user error message
The processTxnError helper function now better distinguishes between metamask exceptions on localhost and Ropsten. The eror message from Metamask in these two cases is contained in a different data structure. Previously, in the Ropsten case, the full exception data was presented to the user as the error message which led to poorly formatted and confusin g error messages.